moved update_defense_positions to match_defenders_for_sim
